backend/ingest-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
pydantic>=2.10.0
aiofiles>=23.2.1
python-multipart>=0.0.9
aiohttp>=3.9.0
